refactor(orchestrator): extract batch and security services to operation layers
